+++ src/lib/newsletter/OptInBanner.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import SubscribeForm from './SubscribeForm.svelte';
+	import { newsletterBannerDescription, newsletterBannerTitle } from './copy';
 
 	type Props = {
 		title?: string;
@@ -14,14 +15,14 @@
 		{#if title}
 			{title}
 		{:else}
-			Get notified about the best new vaults and DeFi news
+			{newsletterBannerTitle}
 		{/if}
 	</h2>
 	<p>
 		{#if description}
 			{description}
 		{:else}
-			Subscribe to Trading Strategy newsletter for updates on DeFi trading strategies and vaults.
+			{newsletterBannerDescription}
 		{/if}
 	</p>
 	<SubscribeForm />